+++ services/bitcore/bitcore.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-sudo chown -Rvf app /app/bitcore
+sudo chown -Rvf app /app/bitcore /app/data/bitcore
 sleep 5
 while true
 do node /app/apps/bitcore/server.js | tee server.log