+++ services/build-one.sh
@@ -1,1 +1,1 @@
-cat 2-build.sh | grep ${1:?which build} | bash
+cat build.sh | grep ${1:?which build} | bash