+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.12
+
+* cleanup scripts
+* juypyter external folder
+*5 volumized data for bitcore, kafka and cassandra
+
 # 0.0.11
 
 - replaced zeppelin with juypter